+++ cpp/microbench/workloads/args_generators/args_generator_builder.h
@@ -8,7 +8,7 @@
 #include "random_xoshiro256p.h"
 #include "json/single_include/nlohmann/json.hpp"
 
-using K = long long;
+using K = int64_t;
 
 // template<typename K>
 struct ArgsGeneratorBuilder {
